editor: update .editorconfig to match new dfmt dlang code style and general codestyle
- `max_line_length = off`:
This formatting process will not keep lines below 80 characters of length. This
makes hard line wrapping not strictly mandatory, but it's recommended 80 to 120
characters of soft wrapping, when possible. For now its disabled due to lack of
support on dfmt.

- `dfmt_brace_style = allman`:
This style puts the brace associated with a control statement on the next line,
indented to the same level as the control statement. Statements within the
braces are indented to the next level.

- `dfmt_soft_max_line_length = 120`:
The formatting process will usually keep lines below 120 characters of length.
Because `max_line_length` property is turned off, there's no hard line wrapping
and this only give you a warning on the editor while editing dlang files.

- `dfmt_align_switch_statements = true`:
Align labels, cases, and defaults with their enclosing switch.

- `dfmt_outdent_attributes = true`: Decrease the indentation level of attributes.

- `dfmt_split_operator_at_line_end = true`:
Place operators on the end of the previous line when splitting lines.

- `dfmt_space_after_cast = true`:
Insert space after the closing paren of a cast expression.

- `dfmt_space_after_keywords = true`:
Insert space after if, while, foreach, etc, and before the (.

- `dfmt_space_before_function_parameters = false`:
Insert space before the opening paren of a function parameter list.

- `dfmt_selective_import_space = true`:
Insert space after the module name and before the : for selective imports.

- `dfmt_compact_labeled_statements = true`:
Place labels on the same line as the labeled switch, for, foreach, or while
statement.

- `dfmt_template_constraint_style = conditional_newline_indent`:
Control the formatting of template constraints.

- `dfmt_single_template_constraint_indent = false`:
Set if the constraints are indented by a single tab instead of two. Has only
an effect if the style set to `always_newline_indent` or
`conditional_newline_indent`.

- `dfmt_space_before_aa_colon = true`:
Adds a space after an associative array key before the : like in older dfmt
versions.

- `dfmt_keep_line_breaks = true`:
Keep existing line breaks if these don't violate other formatting rules.

Any more information about dfmt or editorconfig could be found here:
- https://github.com/dlang-community/dfmt
- https://editorconfig.org/

Signed-off-by: Luís Ferreira <contact@lsferreira.net>